rename _device.py -> _table.py
