+++ pyatmo/weather/_table.py
